+++ src/constraint_system/boolean.rs
@@ -15,28 +15,35 @@ impl TurboComposer {
     /// Note that using this constraint with whatever [`Witness`] that
     /// is not representing a value equalling 0 or 1, will always force the
     /// equation to fail.
-    pub fn gate_boolean(&mut self, a: Witness) {
-        let zero = self.constant_zero();
-        let constraint = Constraint::new().mul(1).output(-BlsScalar::one());
+    pub fn component_boolean(&mut self, a: Witness) {
+        let zero = Self::constant_zero();
+        let constraint = Constraint::new()
+            .mult(1)
+            .output(-BlsScalar::one())
+            .a(a)
+            .b(a)
+            .o(a)
+            .d(zero);
 
-        self.append_gate(a, a, a, zero, constraint);
+        self.append_gate(constraint);
     }
 }
 
 #[cfg(feature = "std")]
 #[cfg(test)]
 mod tests {
-    use super::super::helper::*;
+    use crate::constraint_system::{helper, TurboComposer};
     use dusk_bls12_381::BlsScalar;
+
     #[test]
     fn test_correct_bool_gate() {
-        let res = gadget_tester(
+        let res = helper::gadget_tester(
             |composer| {
-                let zero = composer.constant_zero();
+                let zero = TurboComposer::constant_zero();
                 let one = composer.append_witness(BlsScalar::one());
 
-                composer.gate_boolean(zero);
-                composer.gate_boolean(one);
+                composer.component_boolean(zero);
+                composer.component_boolean(one);
             },
             32,
         );
@@ -45,13 +52,13 @@ mod tests {
 
     #[test]
     fn test_incorrect_bool_gate() {
-        let res = gadget_tester(
+        let res = helper::gadget_tester(
             |composer| {
                 let zero = composer.append_witness(BlsScalar::from(5));
                 let one = composer.append_witness(BlsScalar::one());
 
-                composer.gate_boolean(zero);
-                composer.gate_boolean(one);
+                composer.component_boolean(zero);
+                composer.component_boolean(one);
             },
             32,
         );